+++ components/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @swissjs/components
 
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @swissjs/core@1.2.13
+
+## 0.1.11
+
+### Patch Changes
+
+- Updated dependencies [457e785]
+  - @swissjs/core@1.2.12
+
 ## 0.1.10
 
 ### Patch Changes